feat: moves systray to service and graceful kill
