+++ apps/server/src/main.ts
@@ -7,11 +7,10 @@ import { WsAdapter } from '@nestjs/platform-ws';
 import { ConfigService } from '@nestjs/config';
 import { NestFactory } from '@nestjs/core';
 import { address } from 'ip';
-import SysTray, { MenuItem } from 'systray2';
-import { join } from 'path';
 import { platform } from 'os';
-import { hideConsole, showConsole } from 'node-hide-console-window';
-import open = require('open')
+import { hideConsole } from 'node-hide-console-window';
+import { UtilitiesModule } from './utilities/utilities.module';
+import { SysTrayService } from './utilities/systray.service';
 import { AppModule } from './app.module';
 
 declare const module: any;
@@ -59,7 +58,7 @@ async function bootstrap() {
     const logger = app.get(WINSTON_MODULE_NEST_PROVIDER);
     logger.log(`FlyByWire SimBridge started on: http://${address()}:${port}`, 'NestApplication');
 
-    buildSysTray(logger, isConsoleHidden, port);
+    app.select(UtilitiesModule).get(SysTrayService).init(isConsoleHidden, port);
 
     if (platform() === 'win32' && isConsoleHidden) {
         hideConsole();
@@ -77,84 +76,5 @@ function generateResourceFolders() {
         access(dir, (error) => {
             if (error) mkdirSync(dir, { recursive: true });
         });
-    });
-}
-
-// for TS
-interface MenuItemClickable extends MenuItem {
-    click?: () => void
-}
-
-function buildSysTray(logger, isConsoleHidden: Boolean, port) {
-    let hidden = isConsoleHidden;
-
-    const manageConsole = () => {
-        if (hidden) showConsole();
-        else hideConsole();
-        hidden = !hidden;
-    };
-
-    const remoteDisplayItem = {
-        title: 'Remote Displays',
-        tooltip: 'Open remote displays',
-        items: [{
-            title: 'Open MCDU',
-            tooltip: 'Open the MCDU remote display with your default browser',
-            enabled: true,
-            click: () => {
-                open(`http://localhost:${port}/interfaces/mcdu`);
-            },
-        }],
-    };
-
-    const resourcesFolderItem = {
-        title: 'Open Resources Folder',
-        tooltip: 'Open resource folder in your file explorer',
-        enabled: true,
-        click: () => {
-            open.openApp('explorer', { arguments: [`${process.cwd()}\\resources`] });
-        },
-    };
-
-    const exitItem : MenuItemClickable = {
-        title: 'Exit',
-        tooltip: 'Kill the server',
-        checked: false,
-        enabled: true,
-        click: () => {
-            logger.log('Exiting via Tray', 'Systems Tray');
-            sysTray.kill(true);
-        },
-    };
-
-    const consoleVisibleItem : MenuItemClickable = {
-        title: 'Show/Hide',
-        tooltip: 'Change console visibility',
-        checked: false,
-        enabled: true,
-        click: () => manageConsole(),
-    };
-
-    const sysTray = new SysTray({
-        menu: {
-            icon: join(__dirname, '/assets/images/tail.ico'),
-            title: 'FBW SimBridge',
-            tooltip: 'Flybywire SimBridge',
-            items: [
-                remoteDisplayItem,
-                resourcesFolderItem,
-                consoleVisibleItem,
-                exitItem,
-            ],
-        },
-        copyDir: false,
-    });
-
-    sysTray.onClick((action) => {
-        // eslint-disable-next-line no-prototype-builtins
-        if (action.item.hasOwnProperty('click')) {
-            const item = action.item as MenuItemClickable;
-            item.click();
-        }
     });
 }
