+++ apps/server/src/utilities/utilities.module.ts
@@ -3,10 +3,11 @@ import { UtiliyController } from './utilities.controller';
 import { FileService } from './file.service';
 import { PrinterService } from './printer.service';
 import { ShutDownService } from './shutdown.service';
+import { SysTrayService } from './systray.service';
 
 @Module({
     controllers: [UtiliyController],
-    providers: [FileService, PrinterService, ShutDownService],
+    providers: [FileService, PrinterService, ShutDownService, SysTrayService],
     exports: [FileService, PrinterService],
 })
 export class UtilitiesModule {}
